+++ Assets/RoDOTS/Runtime/UI/UIAttributes.cs
@@ -1,0 +1,6 @@
+﻿using UnityEngine;
+
+namespace RoDOTS.runtime {
+    [CreateAssetMenu(fileName = "UI", menuName = "RoDOTS/UI Data")]
+    public class UIAttributes : UIManager.PoolingAttributes { }
+}